+++ hopsworksai/resource_cluster.go
@@ -28,7 +28,7 @@ func defaultRonDBConfiguration(cloud api.CloudProvider) api.RonDBConfiguration {
 	ronDB := api.RonDBConfiguration{
 		Configuration: api.RonDBBaseConfiguration{
 			NdbdDefault: api.RonDBNdbdDefaultConfiguration{
-				ReplicationFactor: 2,
+				ReplicationFactor: 1,
 			},
 			General: api.RonDBGeneralConfiguration{
 				Benchmark: api.RonDBBenchmarkConfiguration{
@@ -46,7 +46,7 @@ func defaultRonDBConfiguration(cloud api.CloudProvider) api.RonDBConfiguration {
 			NodeConfiguration: api.NodeConfiguration{
 				DiskSize: 512,
 			},
-			Count: 2,
+			Count: 1,
 		},
 		MYSQLNodes: api.WorkerConfiguration{
 			NodeConfiguration: api.NodeConfiguration{
@@ -456,7 +456,7 @@ func ronDBSchema() *schema.Resource {
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"replication_factor": {
-										Description: "The number of replicas created by RonDB for high availability.",
+										Description: "The number of replicas created by RonDB. Set > 1 for high availability.",
 										Type:        schema.TypeInt,
 										Optional:    true,
 										ForceNew:    true,