+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 0.11.0 (Unreleased)
 
 NOTES:
+* resource/hopsworksai_cluster: Change default values for RonDB cluster. New default number of replicas is `1` and new default number of Datanodes is `1`
 
 BREAKING CHANGES:
-
+* Default values for RonDB cluster changed. New default number of replicas is `1` and new default number of Datanodes is `1`
 BUG FIXES:
 
 ENHANCEMENTS: